session updated
completed:
- Database set up for pricing
- session created from Menu -> confirm_order with subtotal pricing

to-do:
- when returning back to menu, display cart items on menu and cart
- add to cart button to NOT refresh the page
- upon order confirmation, post to Database

+++ Main/menu.js
@@ -46,12 +46,15 @@ var choice_5 = document.getElementById("choice_5").innerText;
 var choice_6 = document.getElementById("choice_6").innerText;
 
 // Menu item names
-var price_1 = parseInt(document.getElementById("price_1").innerText.split('$')[1]);
-var price_2 = parseInt(document.getElementById("price_2").innerText.split('$')[1]);
-var price_3 = parseInt(document.getElementById("price_3").innerText.split('$')[1]);
-var price_4 = parseInt(document.getElementById("price_4").innerText.split('$')[1]);
-var price_5 = parseInt(document.getElementById("price_5").innerText.split('$')[1]);
-var price_6 = parseInt(document.getElementById("price_6").innerText.split('$')[1]);
+var price_1 = parseFloat(document.getElementById("price_1").innerText.split('$')[1]);
+var price_2 = parseFloat(document.getElementById("price_2").innerText.split('$')[1]);
+var price_3 = parseFloat(document.getElementById("price_3").innerText.split('$')[1]);
+var price_4 = parseFloat(document.getElementById("price_4").innerText.split('$')[1]);
+var price_5 = parseFloat(document.getElementById("price_5").innerText.split('$')[1]);
+var price_6 = parseFloat(document.getElementById("price_6").innerText.split('$')[1]);
+
+// var item1MenuPrc = document.getElementById("price_1");
+// console.log(parseFloat(item1MenuPrc.innerHTML.split('$')[1]));
 
 // Tabulation DOMs
 var tabulated_sub_total = document.getElementById("tabulated_sub_total");
